+++ actix-web-httpauth/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Implement `From<Basic>` for `BasicAuth`.
 - Minimum supported Rust version (MSRV) is now 1.68.
 
 ## 0.8.0